requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework==3.9.1
whitenoise==4.1.2
django-modeladmin-reorder
python-dateutil